Added base state and PSO setup for testing ACC. Improvements to multi-threading. Changed what state is rendered out in Texture feedback pass.

+++ Editor/EditorBase.h
@@ -20,7 +20,7 @@
 
 inline ImTextureID TextreHandleToIM(FlexKit::ResourceHandle texture)
 {
-    return reinterpret_cast<ImTextureID>(texture.INDEX);
+    return reinterpret_cast<ImTextureID>((size_t)texture.INDEX);
 }
